requirements.txt
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
sqlalchemy>=1.4.0
pymysql>=1.0.0
openai>=1.0.0
python-dotenv>=0.19.0
apscheduler>=3.10.4
pydantic[email]>=1.8.0
pytz>=2023.3
jinja2>=3.0.0
python-multipart>=0.0.5
starlette>=0.14.0